split timeline_track into smaller components
